% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConfidenceIntervals.R
\name{ConfidenceIntervals}
\alias{ConfidenceIntervals}
\title{Creates a new html widget which wraps a chartjs chart instance.}
\usage{
ConfidenceIntervals(
  insertion_type = c("PointForecastFirst", "PointForecastFirst", "OnlyIntervals")[1],
  chart_data,
  top_band,
  bottom_band,
  labels = NULL,
  width = NULL,
  height = NULL,
  elementId = NULL,
  color_main = NULL,
  color_top = NULL,
  color_bottom = NULL,
  show_legend = TRUE,
  show_tooltip = TRUE,
  background_color = "#000000",
  axis_limits = NULL
)
}
\arguments{
\item{insertion_type}{How the confidence intervals are constructed}

\item{chart_data}{The data for the point forecast}

\item{top_band}{The data for the upper bound}

\item{bottom_band}{The data for the lower bound}

\item{labels}{the labels for the data. Displayed on the x-axis.}

\item{width}{width of the widget}

\item{height}{heigth of the widget}

\item{elementId}{The html id of the chart.}
}
\description{
Creates a new html widget which wraps a chartjs chart instance.
}
