% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MyCharts.R
\name{MyCharts}
\alias{MyCharts}
\title{<Add Title>}
\usage{
MyCharts(
  insertion_type = "PointForecastFirst",
  chart_data,
  top_band,
  bottom_band,
  labels,
  width = NULL,
  height = NULL,
  elementId = NULL
)
}
\description{
<Add Description>
}
